Refactor auth forms and pages
