+++ app/src/components/common/SignUpForm.tsx
@@ -5,7 +5,6 @@ import {
   OutlinedInput,
   FormHelperText,
   Stack,
-  Typography,
 } from '@mui/material';
 import {styled} from '@mui/system';
 import GoogleIcon from '@mui/icons-material/Google';
@@ -44,39 +43,37 @@ export const SignUpForm = ({onSubmit}: SignUpFormProps) => {
   });
 
   return (
-    <FormContainer>
-      <FormHeader variant="h4">Sign up for an account</FormHeader>
-      <Form onSubmit={handleSubmit}>
-        <Stack spacing={1}>
-          <InputLabel htmlFor="email">Email address</InputLabel>
-          <OutlinedInput
-            fullWidth
-            id="email"
-            value={values.email}
-            onChange={handleChange}
-            error={touched.email && Boolean(errors.email)}
-          />
-          {touched.email && errors.email && (
-            <FormHelperText error>{errors.email}</FormHelperText>
-          )}
-        </Stack>
-        <Stack spacing={1}>
-          <InputLabel htmlFor="password">Password</InputLabel>
-          <OutlinedInput
-            fullWidth
-            id="password"
-            value={values.password}
-            onChange={handleChange}
-            error={touched.password && Boolean(errors.password)}
-          />
-          {touched.password && errors.password && (
-            <FormHelperText error>{errors.password}</FormHelperText>
-          )}
-        </Stack>
-        <SubmitButton type="submit" fullWidth>
-          Sign Up
-        </SubmitButton>
-      </Form>
+    <Form onSubmit={handleSubmit}>
+      <Stack spacing={1}>
+        <InputLabel htmlFor="email">Email address</InputLabel>
+        <OutlinedInput
+          fullWidth
+          id="email"
+          value={values.email}
+          onChange={handleChange}
+          error={touched.email && Boolean(errors.email)}
+        />
+        {touched.email && errors.email && (
+          <FormHelperText error>{errors.email}</FormHelperText>
+        )}
+      </Stack>
+      <Stack spacing={1}>
+        <InputLabel htmlFor="password">Password</InputLabel>
+        <OutlinedInput
+          fullWidth
+          id="password"
+          type="password"
+          value={values.password}
+          onChange={handleChange}
+          error={touched.password && Boolean(errors.password)}
+        />
+        {touched.password && errors.password && (
+          <FormHelperText error>{errors.password}</FormHelperText>
+        )}
+      </Stack>
+      <SubmitButton type="submit" fullWidth>
+        Sign Up
+      </SubmitButton>
       <Divider>
         <div />
         <p>or</p>
@@ -92,25 +89,9 @@ export const SignUpForm = ({onSubmit}: SignUpFormProps) => {
       >
         <GoogleIcon /> Sign up with Google
       </SocialSignInLink>
-    </FormContainer>
+    </Form>
   );
 };
-
-const FormContainer = styled(Stack)(({theme}) => ({
-  maxWidth: '550px',
-  alignItems: 'center',
-  padding: '2rem',
-  border: '1px solid #e0e0e0',
-  borderRadius: theme.shape.borderRadius,
-  backgroundColor: '#fff',
-  margin: '0 16px',
-}));
-
-const FormHeader = styled(Typography)({
-  textAlign: 'center',
-  marginBottom: '16px',
-  fontWeight: 600,
-});
 
 const Form = styled('form')({
   width: '100%',
@@ -118,7 +99,6 @@ const Form = styled('form')({
   flexDirection: 'column',
   alignItems: 'stretch',
   gap: '1rem',
-  marginBottom: '16px',
 });
 
 const SubmitButton = styled(PrimaryButton)({
@@ -127,7 +107,6 @@ const SubmitButton = styled(PrimaryButton)({
 
 const Divider = styled(Box)(({theme}) => ({
   color: theme.palette.text.secondary,
-  marginBottom: '16px',
   display: 'flex',
   alignItems: 'center',
   width: '100%',
